User/requirements.txt
boto3
streamlit
python-dotenv
sqlalchemy
